requirements.txt
supabase==2.3.4
streamlit==1.31.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0